DCJ25-11: Created DiceRollManager
Additionally, I have created a new DungeonCrawlerPawn class to be the parent of Enemy, Player, and Chicken. I'm keeping the Actor class to work with LineTracing further down the road. But the Pawn is so that I can have a parent that will pass along the Components to the Enemy, Player and Chicken child classes